MB-15165: Add option to load data from file
It's now possible to load data from file. Currently only a file format
for multidimensional data is supported. The file is just a binary file
containing 64-bit floats. There's a rectangle given for every dimension.
You need to supply the dimensionality of the input file in order to
have it work correctly, as the dimensionality can't be determined from
the file itself.

Change-Id: I4459b6bbf3a433098781eb90d59df77b328ee3c2
Reviewed-on: http://review.couchbase.org/51632
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Volker Mische <volker.mische@gmail.com>