Introduce `index_mode`
As perfrunner can run tests for several different indexes (mapreduce views,
spatial views and n1ql), it makes sense to have a setting to define which
one should be used. This is what the `index_mode` is for. It can be
`mapreduce` (the default) or `n1ql`. In the future also support for `spatial`
will be added.

This change also needs a change in perfrunner that sets the index mode.