Add support for running spatial queries from file
It's now possible to run queries from a file. Currently only a file format
for multidimensional data is supported. The file is just a binary file
containing 64-bit floats. There's a rectangle given for every dimension.
You need to supply the dimensionality of the input file in order to
have it work correctly, as the dimensionality can't be determined from
the file itself.