Append Old to class names used for N1QL
This changes NewN1QLQuery to OldN1QLQuery and N1QLWorker to
OldN1QLWorker. We will be adding in new N1QL functionality, but
don't want to risk breaking any of the current code. In a future
change we will remove these classes and convert the old tests to
use thenew code.