Add capped fields so query results are fixed
This chanage adds two fields that are guarneteed to contain a
fixed number of similar data across an entire data set. The two
fields are capped_small and capped_large and they will contain
100 and 3000 similar values respectively across the dataset.

Note that for this capped function to work properly document keys
must have an increasing number appended to the end of the key.