Allow for n1ql queries to work with a partitioned key space
We need to partition the key space when running n1ql queries because
we cannot update fields for keys that we are actually querying. In
order to do this we have a special 'n1ql' prefixed key space that
we use for querying. The default key space is then used for updates.
This allows us to update the index while still querying it at a
high rate.

Change-Id: I8150717dd6d34f07a979a987e274d09cbe15b388
Reviewed-on: http://review.couchbase.org/54189
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>